Created folder for Failure Types, moved the existing one into there and updated namespaces
